requirements.txt
allure-pytest
selenium
pytest